refactor(#306): 패키지 분리
