+++ AOS/app/src/main/java/boostcamp/and07/mindsync/data/network/MindMapSocketManager.kt
@@ -2,8 +2,8 @@ package boostcamp.and07.mindsync.data.network
 
 import android.util.Log
 import boostcamp.and07.mindsync.BuildConfig
-import boostcamp.and07.mindsync.data.network.request.JoinBoard
-import boostcamp.and07.mindsync.data.network.request.UpdateMindMap
+import boostcamp.and07.mindsync.data.network.request.socket.JoinBoardRequest
+import boostcamp.and07.mindsync.data.network.request.socket.UpdateMindMapRequest
 import boostcamp.and07.mindsync.data.network.response.mindmap.SerializedCrdtTree
 import boostcamp.and07.mindsync.data.network.response.mindmap.SerializedOperation
 import io.socket.client.IO
@@ -73,7 +73,7 @@ class MindMapSocketManager {
         boardId: String,
         boardName: String,
     ) {
-        socket.emit(EVENT_JOIN_BOARD, Json.encodeToString(JoinBoard(boardId, boardName)))
+        socket.emit(EVENT_JOIN_BOARD, Json.encodeToString(JoinBoardRequest(boardId, boardName)))
     }
 
     fun updateMindMap(
@@ -82,7 +82,7 @@ class MindMapSocketManager {
     ) {
         socket.emit(
             EVENT_UPDATE_MIND_MAP,
-            Json.encodeToString(UpdateMindMap(serializedOperation, boardId)),
+            Json.encodeToString(UpdateMindMapRequest(serializedOperation, boardId)),
         )
     }
