+++ AOS/app/src/main/java/boostcamp/and07/mindsync/data/network/api/LoginApi.kt
@@ -1,7 +1,7 @@
-package boostcamp.and07.mindsync.data.network
+package boostcamp.and07.mindsync.data.network.api
 
-import boostcamp.and07.mindsync.data.network.request.KakaoLoginRequest
-import boostcamp.and07.mindsync.data.network.response.LoginResponse
+import boostcamp.and07.mindsync.data.network.request.login.KakaoLoginRequest
+import boostcamp.and07.mindsync.data.network.response.login.LoginResponse
 import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.POST
