+++ AOS/app/src/main/java/boostcamp/and07/mindsync/data/network/api/LogoutApi.kt
@@ -1,6 +1,6 @@
-package boostcamp.and07.mindsync.data.network
+package boostcamp.and07.mindsync.data.network.api
 
-import boostcamp.and07.mindsync.data.network.request.LogoutRequest
+import boostcamp.and07.mindsync.data.network.request.login.LogoutRequest
 import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.POST
