+++ AOS/app/src/main/java/boostcamp/and07/mindsync/data/network/api/TokenApi.kt
@@ -1,7 +1,7 @@
-package boostcamp.and07.mindsync.data.network
+package boostcamp.and07.mindsync.data.network.api
 
-import boostcamp.and07.mindsync.data.network.request.NewAccessTokenRequest
-import boostcamp.and07.mindsync.data.network.response.NewAccessTokenResponse
+import boostcamp.and07.mindsync.data.network.request.login.NewAccessTokenRequest
+import boostcamp.and07.mindsync.data.network.response.login.NewAccessTokenResponse
 import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.POST
