+++ AOS/app/src/main/java/boostcamp/and07/mindsync/data/network/request/login/NewAccessTokenRequest.kt
@@ -1,12 +1,10 @@
-package boostcamp.and07.mindsync.data.network.request
+package boostcamp.and07.mindsync.data.network.request.login
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class LogoutRequest(
+data class NewAccessTokenRequest(
     @SerialName("refresh_token")
     val refreshToken: String,
-    @SerialName("access_token")
-    val accessToken: String,
 )
