+++ AOS/app/src/main/java/boostcamp/and07/mindsync/data/network/request/socket/UpdateMindMapRequest.kt
@@ -1,10 +1,10 @@
-package boostcamp.and07.mindsync.data.network.response
+package boostcamp.and07.mindsync.data.network.request.socket
 
 import boostcamp.and07.mindsync.data.network.response.mindmap.SerializedOperation
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class OperationFromServer(
+data class UpdateMindMapRequest(
     val operation: SerializedOperation,
     val boardId: String,
 )
