+++ AOS/app/src/main/java/boostcamp/and07/mindsync/data/network/response/board/CreateBoardDto.kt
@@ -2,9 +2,10 @@ package boostcamp.and07.mindsync.data.network.response.board
 
 import kotlinx.serialization.Serializable
 
+
 @Serializable
-data class BoardsResponse(
-    val statusCode: Int,
-    val message: String,
-    val data: List<BoardDto>,
+data class CreateBoardData(
+    val boardId: String,
+    val date: String,
+    val imageUrl: String,
 )
