+++ AOS/app/src/main/java/boostcamp/and07/mindsync/data/network/response/board/CreateBoardResponse.kt
@@ -9,9 +9,3 @@ data class CreateBoardResponse(
     val data: CreateBoardData?,
 )
 
-@Serializable
-data class CreateBoardData(
-    val boardId: String,
-    val date: String,
-    val imageUrl: String,
-)
