+++ AOS/app/src/main/java/boostcamp/and07/mindsync/data/network/response/login/LoginDto.kt
@@ -1,10 +1,13 @@
-package boostcamp.and07.mindsync.data.network.request
+package boostcamp.and07.mindsync.data.network.response.login
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
+
 @Serializable
-data class NewAccessTokenRequest(
+data class LoginDto(
+    @SerialName("access_token")
+    val accessToken: String,
     @SerialName("refresh_token")
     val refreshToken: String,
 )
