+++ AOS/app/src/main/java/boostcamp/and07/mindsync/data/network/response/login/LoginResponse.kt
@@ -1,12 +1,13 @@
-package boostcamp.and07.mindsync.data.network.response.user
+package boostcamp.and07.mindsync.data.network.response.login
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class UserResponse(
+data class LoginResponse(
     val statusCode: Int,
     val message: String,
-    val data: List<UserDto>? = null,
+    val error: String? = null,
+    val data: LoginDto? = null,
 )
