+++ AOS/app/src/main/java/boostcamp/and07/mindsync/data/network/response/login/NewAccessTokenResponse.kt
@@ -1,4 +1,4 @@
-package boostcamp.and07.mindsync.data.network.response
+package boostcamp.and07.mindsync.data.network.response.login
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
@@ -8,11 +8,6 @@ data class NewAccessTokenResponse(
     val statusCode: Int,
     val message: String,
     val error: String? = null,
-    val data: NewAccessTokenData? = null,
+    val data: NewAccessTokenDto? = null,
 )
 
-@Serializable
-data class NewAccessTokenData(
-    @SerialName("access_token")
-    val accessToken: String,
-)
