+++ AOS/app/src/main/java/boostcamp/and07/mindsync/data/network/response/mindmap/SerializedCrdtTree.kt
@@ -13,37 +13,3 @@ data class SerializedCrdtTree(
     @SerialName("tree")
     val tree: SerializedTree,
 )
-
-@Serializable
-data class SerializedOperationLog(
-    val operation: SerializedOperation,
-    val oldDescription: String? = null,
-    val oldParentId: String? = null,
-)
-
-@Serializable
-data class SerializedOperation(
-    val operationType: String,
-    val id: String,
-    val clock: Clock,
-    val description: String? = null,
-    val parentId: String? = null,
-)
-
-@Serializable
-data class SerializedTree(
-    @SerialName("nodes")
-    val nodes: List<NodeDto>,
-)
-
-@Serializable
-data class NodeDto(
-    @SerialName("children")
-    val children: List<String>,
-    @SerialName("targetId")
-    val targetId: String,
-    @SerialName("parentId")
-    val parentId: String,
-    @SerialName("description")
-    val description: String = "",
-)
