+++ AOS/app/src/main/java/boostcamp/and07/mindsync/data/network/response/mindmap/SerializedOperationLog.kt
@@ -1,0 +1,10 @@
+package boostcamp.and07.mindsync.data.network.response.mindmap
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class SerializedOperationLog(
+    val operation: SerializedOperation,
+    val oldDescription: String? = null,
+    val oldParentId: String? = null,
+)
