+++ AOS/app/src/main/java/boostcamp/and07/mindsync/data/network/response/profile/ProfileResponse.kt
@@ -1,4 +1,4 @@
-package boostcamp.and07.mindsync.data.network.response
+package boostcamp.and07.mindsync.data.network.response.profile
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
