+++ AOS/app/src/main/java/boostcamp/and07/mindsync/data/network/response/socket/OperationFromServer.kt
@@ -1,10 +1,10 @@
-package boostcamp.and07.mindsync.data.network.request
+package boostcamp.and07.mindsync.data.network.response.socket
 
 import boostcamp.and07.mindsync.data.network.response.mindmap.SerializedOperation
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class UpdateMindMap(
+data class OperationFromServer(
     val operation: SerializedOperation,
     val boardId: String,
 )
