+++ AOS/app/src/main/java/boostcamp/and07/mindsync/data/repository/login/LoginRepository.kt
@@ -1,7 +1,7 @@
 package boostcamp.and07.mindsync.data.repository.login
 
-import boostcamp.and07.mindsync.data.network.response.LoginData
+import boostcamp.and07.mindsync.data.network.response.login.LoginDto
 
 interface LoginRepository {
-    suspend fun loginWithKakao(kakaoUserId: Long): Result<LoginData>
+    suspend fun loginWithKakao(kakaoUserId: Long): Result<LoginDto>
 }
