+++ AOS/app/src/main/java/boostcamp/and07/mindsync/data/repository/login/LoginRepositoryImpl.kt
@@ -1,15 +1,15 @@
 package boostcamp.and07.mindsync.data.repository.login
 
-import boostcamp.and07.mindsync.data.network.LoginApi
-import boostcamp.and07.mindsync.data.network.request.KakaoLoginRequest
-import boostcamp.and07.mindsync.data.network.response.LoginData
+import boostcamp.and07.mindsync.data.network.api.LoginApi
+import boostcamp.and07.mindsync.data.network.request.login.KakaoLoginRequest
+import boostcamp.and07.mindsync.data.network.response.login.LoginDto
 import boostcamp.and07.mindsync.ui.util.NetworkExceptionMessage
 import javax.inject.Inject
 
 class LoginRepositoryImpl
     @Inject
     constructor(private val api: LoginApi) : LoginRepository {
-        override suspend fun loginWithKakao(kakaoUserId: Long): Result<LoginData> {
+        override suspend fun loginWithKakao(kakaoUserId: Long): Result<LoginDto> {
             return try {
                 val request = KakaoLoginRequest(kakaoUserId)
                 val response = api.postKakaoOAuth(request)
