+++ AOS/app/src/main/java/boostcamp/and07/mindsync/data/repository/profilespace/ProfileSpaceRepository.kt
@@ -2,7 +2,7 @@ package boostcamp.and07.mindsync.data.repository.profilespace
 
 import boostcamp.and07.mindsync.data.model.Space
 import boostcamp.and07.mindsync.data.network.response.space.ProfileSpaceJoinData
-import boostcamp.and07.mindsync.data.network.response.user.UserData
+import boostcamp.and07.mindsync.data.network.response.user.UserDto
 import kotlinx.coroutines.flow.Flow
 
 interface ProfileSpaceRepository {
@@ -10,7 +10,7 @@ interface ProfileSpaceRepository {
 
     fun getSpaces(): Flow<List<Space>>
 
-    fun getSpaceUsers(spaceUuid: String): Flow<List<UserData>>
+    fun getSpaceUsers(spaceUuid: String): Flow<List<UserDto>>
 
     fun leaveSpace(spaceId: String): Flow<Boolean>
 }
