+++ AOS/app/src/main/java/boostcamp/and07/mindsync/data/repository/profilespace/ProfileSpaceRepositoryImpl.kt
@@ -1,10 +1,10 @@
 package boostcamp.and07.mindsync.data.repository.profilespace
 
 import boostcamp.and07.mindsync.data.model.Space
-import boostcamp.and07.mindsync.data.network.ProfileSpaceApi
-import boostcamp.and07.mindsync.data.network.request.ProfileSpaceJoinRequest
+import boostcamp.and07.mindsync.data.network.api.ProfileSpaceApi
+import boostcamp.and07.mindsync.data.network.request.space.ProfileSpaceJoinRequest
 import boostcamp.and07.mindsync.data.network.response.space.ProfileSpaceJoinData
-import boostcamp.and07.mindsync.data.network.response.user.UserData
+import boostcamp.and07.mindsync.data.network.response.user.UserDto
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 import javax.inject.Inject
@@ -40,7 +40,7 @@ class ProfileSpaceRepositoryImpl
                 }
             }
 
-        override fun getSpaceUsers(spaceId: String): Flow<List<UserData>> =
+        override fun getSpaceUsers(spaceId: String): Flow<List<UserDto>> =
             flow {
                 val response = profileSpaceApi.getSpaceUsers(spaceId)
                 response.data?.let { users ->
