+++ AOS/app/src/main/java/boostcamp/and07/mindsync/ui/main/MainBindingAdpater.kt
@@ -3,11 +3,11 @@ package boostcamp.and07.mindsync.ui.main
 import androidx.databinding.BindingAdapter
 import androidx.recyclerview.widget.RecyclerView
 import boostcamp.and07.mindsync.data.model.Space
-import boostcamp.and07.mindsync.data.network.response.user.UserData
+import boostcamp.and07.mindsync.data.network.response.user.UserDto
 import boostcamp.and07.mindsync.ui.boardlist.UsersAdapter
 
 @BindingAdapter("app:users")
-fun RecyclerView.bindUsers(users: List<UserData>) {
+fun RecyclerView.bindUsers(users: List<UserDto>) {
     if (this.adapter != null) {
         (this.adapter as UsersAdapter).submitList(users.toMutableList())
     }
