+++ AOS/app/src/main/java/boostcamp/and07/mindsync/ui/main/MainUiState.kt
@@ -1,11 +1,11 @@
 package boostcamp.and07.mindsync.ui.main
 
 import boostcamp.and07.mindsync.data.model.Space
-import boostcamp.and07.mindsync.data.network.response.user.UserData
+import boostcamp.and07.mindsync.data.network.response.user.UserDto
 
 data class MainUiState(
     val profileImageUrl: String = "",
     val spaces: List<Space> = listOf(),
     val nowSpace: Space? = null,
-    val users: List<UserData> = listOf(),
+    val users: List<UserDto> = listOf(),
 )
