+++ AOS/app/src/main/res/layout/item_users.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="user"
-            type="boostcamp.and07.mindsync.data.network.response.user.UserData" />
+            type="boostcamp.and07.mindsync.data.network.response.user.UserDto" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
